feat: :sparkles: Added SuggestionsList component for displaying city suggestions
I have added the `SuggestionsList` component to display a list of suggested cities based on the user input.
When a user clicks on a suggestion, it updates the input field with the selected city and clears the hint and suggestions list.
The `App` component has been updated to manage and render this new component, which uses a filtered list of city names based on the current input value.